Support Decimal Zoom Steps and Update API
Add support for requesting clusters on decimal zoom steps in the Supercluster JavaScript library. Previously, only integer zoom levels were supported. Introduce a new 'stepSize' parameter and incorporate it into the API.

This update includes internal modifications to the handling of zoom levels, allowing for accommodation of decimal zoom steps. As an API-breaking change, the 'clusterId' structure has been altered. Previously, the 'clusterId' encoded the zoom level within the ID, supporting up to 32 levels. Now, it encodes the index of the 'trees' array.